Challenge: The floor is lava!
It’s time to put your new skills to the test! In this challenge assignment, you will create a track for a ball to roll on without touching the floor. Make your track as long as you dare, but don’t let the ball drop, because the floor is lava.
When you’ve finished, you’ll be able to submit your work, and see what others have submitted.